+++ modules/src/image/fourier/fftzoom.c
@@ -38,8 +38,8 @@ void fftzoom(Fimage in, Fimage out, float *z, char *i_flag)
   ny = in_re->nrow;
   zoom = (i_flag?1./(*z):*z);
   if (zoom<=0.) mwerror(FATAL,1,"Zoom factor must be positive.\n");
-  nxz = (int) floor((double)nx*(double)zoom + .5);
-  nyz = (int) floor((double)ny*(double)zoom + .5);
+  nxz = floor((double)nx*(double)zoom + .5);
+  nyz = floor((double)ny*(double)zoom + .5);
   z2 = zoom*zoom;
   
   /* ALLOCATE AND INITIALIZE IMAGES */
